20160926 - 更新sdk,新增读取设备亮度值

+++ interface_sdk/src/main/java/com/core/gatewayinterface/DataSources.java
@@ -93,7 +93,7 @@ public class DataSources {
     }
 
     //获取设备亮度值
-    public void getDeviceLevel(String deviceId,byte value){
+    public void getDeviceLevel(String deviceId,int value){
         sdkCallback.getDeviceLevelCallback(deviceId,value);
     }
